int to uint

+++ include/system/data.h
@@ -24,8 +24,8 @@ extern void data_dtor(struct _data *this);
 extern void data_push(struct _data *this, const int8_t value);
 extern int8_t data_pop(const struct _data *this);
 
-extern void data_write(struct _data *this, const int addr, const int8_t value);
-extern int8_t data_read(const struct _data *this, const int addr);
+extern void data_write(struct _data *this, const uint16_t addr, const int8_t value);
+extern int8_t data_read(const struct _data *this, const uint16_t addr);
 
 extern void data_coi(const struct _data *this, tuple_t *buffer);
 extern void data_dump(const struct _data *this, array_t *buffer);
